# Speed up REGEX for invalid fields in MiqExpression::Field